requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5